Serving a file on a get HTTP req
In this example I server a file on app.get request. I do this by using the res.sendFile. __dirname is the global variable for the current directory or file path that the user is on. The concatenation with the string is the location of the file that we want to send.